+++ api/controllers/service_api/app/file.py
@@ -4,7 +4,7 @@ from flask_restful import marshal_with
 from controllers.service_api import api
 from controllers.service_api.wraps import AppApiResource
 from controllers.service_api.app import create_or_update_end_user_for_user_id
-from controllers.service_api.app.file_error import NoFileUploadedError, TooManyFilesError, FileTooLargeError, \
+from controllers.service_api.app.error import NoFileUploadedError, TooManyFilesError, FileTooLargeError, \
     UnsupportedFileTypeError
 import services
 from services.file_service import FileService
